+++ reactor-extra/src/main/kotlin/reactor/retry/RetryExtensions.kt
@@ -19,7 +19,7 @@ import java.util.function.Consumer
  * @author Simon Baslé
  * @since 3.1.1
  */
-fun <T> Flux<T>.retryExponentialBackoff(times: Int, first: Duration, max: Duration? = null,
+fun <T> Flux<T>.retryExponentialBackoff(times: Long, first: Duration, max: Duration? = null,
                                         jitter: Boolean = false,
                                         doOnRetry: ((RetryContext<T>) -> Unit)? = null): Flux<T> {
     val retry = Retry.any<T>()
@@ -46,7 +46,7 @@ fun <T> Flux<T>.retryExponentialBackoff(times: Int, first: Duration, max: Durati
  * @author Simon Baslé
  * @since 3.1.1
  */
-fun <T> Flux<T>.retryRandomBackoff(times: Int, first: Duration, max: Duration? = null,
+fun <T> Flux<T>.retryRandomBackoff(times: Long, first: Duration, max: Duration? = null,
                                    doOnRetry: ((RetryContext<T>) -> Unit)? = null): Flux<T> {
     val retry = Retry.any<T>()
             .retryMax(times)
@@ -72,7 +72,7 @@ fun <T> Flux<T>.retryRandomBackoff(times: Int, first: Duration, max: Duration? =
  * @author Simon Baslé
  * @since 3.1.1
  */
-fun <T> Mono<T>.retryExponentialBackoff(times: Int, first: Duration, max: Duration? = null,
+fun <T> Mono<T>.retryExponentialBackoff(times: Long, first: Duration, max: Duration? = null,
                                         jitter: Boolean = false,
                                         doOnRetry: ((RetryContext<T>) -> Unit)? = null): Mono<T> {
     val retry = Retry.any<T>()
@@ -99,7 +99,7 @@ fun <T> Mono<T>.retryExponentialBackoff(times: Int, first: Duration, max: Durati
  * @author Simon Baslé
  * @since 3.1.1
  */
-fun <T> Mono<T>.retryRandomBackoff(times: Int, first: Duration, max: Duration? = null,
+fun <T> Mono<T>.retryRandomBackoff(times: Long, first: Duration, max: Duration? = null,
                                    doOnRetry: ((RetryContext<T>) -> Unit)? = null): Mono<T> {
     val retry = Retry.any<T>()
             .retryMax(times)
